# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LowPriceClone",
        "description": "List of endpoints",
        "contact": {
            "name": "Swagger API Team"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "API server"
        }
    ],
    "paths": {
...
        "/api/order/store": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Create a new order",
                "operationId": "861e98a008b0e483c96ecef6de174a2e",
                "requestBody": {
                    "description": "Order object to be created",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "paymentmethode_id",
                                    "methodelivraison_id",
                                    "addresse_id",
                                    "prixTotal",
                                    "items"
                                ],
                                "properties": {
                                    "paymentmethode_id": {
                                        "type": "integer",
                                        "example": "3"
                                    },
                                    "methodelivraison_id": {
                                        "type": "integer",
                                        "example": "2"
                                    },
                                    "addresse_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "prixTotal": {
                                        "type": "number",
                                        "example": "250000"
                                    },
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "produit_id",
                                                "quantity",
                                                "subTotal"
                                            ],
                                            "properties": {
                                                "produit_id": {
                                                    "type": "integer",
                                                    "example": "28"
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "example": "1"
                                                },
                                                "subTotal": {
                                                    "type": "number",
                                                    "example": "252000"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Order created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Order created successfully"
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "paymentmethode_id": {
                                                    "type": "integer"
                                                },
                                                "methodelivraison_id": {
                                                    "type": "integer"
                                                },
                                                "addresse_id": {
                                                    "type": "integer"
                                                },
                                                "prixTotal": {
                                                    "type": "number"
                                                },
                                                "items": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "produit_id": {
                                                                "type": "integer"
                                                            },
                                                            "quantity": {
                                                                "type": "integer"
                                                            },
                                                            "subTotal": {
                                                                "type": "number"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/order/mescommandes": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get orders of the authenticated user",
                "operationId": "8423a5ba2a0e58526f01c7afb704dc05",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sort_field",
                        "in": "query",
                        "description": "Field to sort by",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort_direction",
                        "in": "query",
                        "description": "Sort direction (asc or desc)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of orders of the user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "prixTotal": {
                                                "type": "string"
                                            },
                                            "orderidentify": {
                                                "type": "string"
                                            },
                                            "first_product_image": {
                                                "type": "string"
                                            },
                                            "statut": {
                                                "type": "string"
                                            },
                                            "created_at": {
                                                "type": "string"
                                            },
                                            "updated_at": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT token should be provided in the 'Authorization' header. Example: Bearer {token}",
                "scheme": "bearer"
            }
        }
    }
}